% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit.hsblock}
\alias{fit.hsblock}
\title{Fit a hierarchical stochastic block on the network data}
\usage{
fit.hsblock(A, distrib = c("bernoulli", "poisson"), Opt = list(),
  Z = NULL, degree.corrected = FALSE, tree.depth = NULL, rseed = NULL,
  vbiter = NULL, inner.iter = NULL, burnin.iter = NULL,
  record.interval = NULL, verbose = TRUE, rate = 0.01, decay = -0.55,
  delay = 1)
}
\arguments{
\item{A}{[n x n] adjacency matrix}

\item{distrib}{type of distribution for edge weights (bernoulli, poisson)}

\item{Opt}{a list of options to feed to set the following options simultaneously}

\item{Z}{[K x n] latent assignment matrix#'}

\item{degree.corrected}{fit degree-corrected model or not (default: FALSE)}

\item{tree.depth}{depth of binary tree model should be greater than 1 (default: 2)}

\item{rseed}{random seed for Monte Carlo steps (default: 19937)}

\item{vbiter}{number of variational Bayes iterations (default: 1000)}

\item{inner.iter}{number of Monte Carlo sampling (default: 100)}

\item{burnin.iter}{number of Monte Carlo sampling to discard (default: 100)}

\item{record.interval}{recording interval (default: 10)}

\item{verbose}{verbosity (default: TRUE)}

\item{rate}{base rate for stochastic gradient update (default: 0.01)}

\item{decay}{rate decay parameter (default: -0.55)}

\item{delay}{rate delay parameter (default: 1)}
}
\value{
a list of variatinal inference results
}
\description{
Fit a hierarchical stochastic block on the network data
}
\examples{

# Build a network from a list of pairs
pairs.file <- system.file('extdata', 'football.pairs', package = 'hsblock')
pairs <- read.table(pairs.file, stringsAsFactors = FALSE)
net.data <- pairs.to.sparse.matrix(pairs)
A <- net.data$A

out <- fit.hsblock(A, distrib = 'bernoulli', tree.depth = 5, vbiter = 1000, rate = 0.01)

library(Matrix)
oo <- order(apply(out$Z, 2, which.max))
image(A[oo, oo])

y.lim <- range(c(out$llik.variational, out$llik.empirical))
plot(out$llik.empirical, ylim = y.lim, type = 'b')
points(out$llik.variational, col = 2, type = 'b', pch = 19, cex = .5)

}
\seealso{
\code{pairs.to.sparse.matrix}

\code{build.random.sparse.Z}
}
\author{
Yongjin Park, \email{ypp@csail.mit.edu}, \email{yongjin.peter.park@gmail.com}
}
\keyword{EM}
\keyword{clustering}
\keyword{network}
