\name{HSBLOCK}
\alias{hsblock-package}
\alias{hsblock}
\docType{package}
\title{
\packageTitle{hsblock}
}
\description{
\packageDescription{hsblock}
}
\details{}
\author{
\packageAuthor{hsblock}
Maintainer: \packageMaintainer{hsblock}
}
\references{Park and Bader}
\keyword{package}
\seealso{}
\examples{}
