% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pairs.to.sparse.matrix}
\alias{pairs.to.sparse.matrix}
\title{Build a sparse matrix from a list pairs or weighted pairs}
\usage{
\method{pairs}{to.sparse.matrix}(pairs, vertices = NULL, default.weight = 1,
  symmetrize = TRUE)
}
\arguments{
\item{pairs}{a data.frame of network edges (m x 2) or (m x 3) where m is number of edges}

\item{vertices}{a set of vertices}

\item{default.weight}{default edge weight (default: 1)}

\item{symmetrize}{build a symmetric adjacency matrix (default: TRUE)}
}
\value{
a list that contains (A = Matrix::sparseMatrix, V = vertices)
}
\description{
Build a sparse matrix from a list pairs or weighted pairs
}
\examples{

pairs.file <- system.file('extdata', 'football.pairs', package = 'hsblock')
pairs <- read.table(pairs.file, stringsAsFactors = FALSE)
net.data <- pairs.to.sparse.matrix(pairs)

library(Matrix)
image(net.data$A[1:10, 1:10])

}
